trival_example: open raw pcm files in binary mode.
The simple codec round-trip example file in the doc directory
opens an input and output pcm file. It was working fine on
POSIX systems, but not on Windows, which treats text files
differently.

This is confusing in a example, so it's better to add an
explicit binary flag to the fopen() calls. This does nothing
on unix-like systems, but should make the example work for
developers on Windows.

Thanks to Wavesonics who reported this on irc.